* Fix components path (#23)

* Fix URLs to load assets & fonts (#24)

* Fix css files glob to replace URLs (#25)

* Update config to build package & test as prod (#26)

* Update components parameters & behavior (#27)

* Verify built files in check workflow (#28)

+++ app/docker-cicd-command.sh
@@ -2,8 +2,11 @@
 
 set -e
 
-# Update the dependencies
-python -m build
+# Assert the dist folder is not empty
+if [ ! "$(ls -A ./dist 2> /dev/null)" ]; then
+  echo "Error: The dist folder is empty. Try to rebuild the image." >&2
+  exit 1
+fi
 
 # Copy the dist folder to /tmp/dist
 rm -rf /tmp/dist/*
